created html_to_str_spec.rb, made first test, failed first test, removed the failing function in html_to_str.rb and replaced with Sring.lstrip

+++ html/html_to_str.rb
@@ -20,18 +20,10 @@
 #   -file_out                     the output file to write to. set to
 #                                 str_out.txt by default
 
-$filename = "../CaptivePortal/test config page.html"
+$filename = "test config page.html"
 $file_out = "str_out.txt"
 keep_shape = true # keep shape of html document? i.e. reinsert the whitespace at the begining of each line?
 $file_buffer = ""
-
-def trim_tail string
-  i = 1
-  while string[-i].match? /\S/
-    i += 1
-  end
-  string[1..-i]
-end
 
 def include_tabs line
   # used to create the initial string for parse_line
@@ -43,7 +35,7 @@ def include_tabs line
 end
 
 def trimmed_and_stringed string
-  return "\"#{trim_tail(string)}\""
+  return "\"#{string.lstrip}\""
 end
 
 def parse_line line
@@ -58,9 +50,11 @@ def parse_line line
 end
 
 def load_file(filename = $filename)
-  File.open(filename, "r") do |html_file|
-    html_file.each do |line|
-      $file_buffer.append(line)
+  if File.exists?(filename)
+    File.open(filename, "r") do |html_file|
+      html_file.each do |line|
+        $file_buffer.append(line)
+      end
     end
   end
 end
